+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+<h4>0.17.1</h4>
+
+* comment.list_all added
+
 <h4>0.17.0</h4>
 
 * app endpoint